carts routes ant controller

+++ app/services/cart_service.ts
@@ -1,24 +1,55 @@
 import Cart from '#models/cart'
 
+//validator
+import { createCartValidator } from '#validators/cart_validator'
+
+//Exceptions
+import HTTPNotFoundException from '#exceptions/HTTP_not_found_exception'
+import HTTPAlreadyExistsException from '#exceptions/HTTP_already_exists_exception'
+
 export class CartService {
-  async getById(id: number) {
-    return await Cart.findOrFail(id)
+  async show(user_id: number) {
+    const hasCart = await Cart.findBy('user_id', user_id)
+
+    if (!hasCart) {
+      const cartData = {
+        user_id: Number(user_id),
+        cart_price: 0.0,
+      }
+
+      const cartPayload = await createCartValidator.validate(cartData)
+
+      const cart = await Cart.create(cartPayload)
+      return cart
+    }
+
+    return hasCart
   }
 
   async create(data: Partial<Cart>) {
+    const hasCart = await Cart.findBy('user_id', data.user_id)
+
+    if (hasCart) throw new HTTPAlreadyExistsException('Cart already exists')
+
     const cart = await Cart.create(data)
     return cart
   }
 
-  async update(id: number, data: Partial<Cart>) {
-    const cart = await Cart.findOrFail(id)
+  async update(user_id: number, data: Partial<Cart>) {
+    const cart = await Cart.findBy('user_id', user_id)
+    if (!cart) throw new HTTPNotFoundException('Cart not found')
+
     cart.merge(data)
     await cart.save()
+
     return cart
   }
 
-  async delete(id: number) {
-    const cart = await Cart.findOrFail(id)
+  async delete(user_id: number) {
+    const cart = await Cart.findBy('user_id', user_id)
+
+    if (!cart) throw new HTTPNotFoundException('Cart not found')
+
     await cart.delete()
     return { message: 'Cart deleted successfully' }
   }
